% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ode_solver.R
\name{cm_ode_c}
\alias{cm_ode_c}
\title{ODE solver}
\usage{
cm_ode_c(dat, state, parms, time, cm, maxsteps = 1e+06, hmax = 0, ...)
}
\arguments{
\item{dat}{a data.frame with ode data}

\item{state}{a vector of initial states}

\item{parms}{a vector of parameters}

\item{cm}{an object of class custommodel}

\item{maxsteps}{maxsteps parameter, defaults to 1e6}

\item{hmax}{hmax parameter, defaults to 0}

\item{parms}{a vector of times}
}
\description{
Takes a data table with the same experiment and solves an ode for it. Parameters & states are considered equal for all data points
}
\examples{
cm_ode_c(dat, state, parms, time, cm)


}
\keyword{ode}

