% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_cm.R
\name{create_cm}
\alias{create_cm}
\title{A custommodel-class constructor}
\usage{
create_cm(data_ode = NULL, data_expl_list = NULL, explfun = NULL,
  cname = NULL, ode_names = NULL, ode_parnames = NULL, fit_start = NULL)
}
\arguments{
\item{data_ode}{a data.frame with ode data or a name of xlsx file with data}

\item{data_expl_list}{a list of data.frames with explicit data or a name of xlsx file with data (each data.frame in a separate sheet)}

\item{explfun}{a list of explicit functions (length equal to length of data_expl_list)}

\item{cname}{a name of the compiled c file to be loaded}

\item{ode_names}{a vector of names of y's for ode system}

\item{ode_parnames}{a vector of names of parameters for ode system}

\item{fit_start}{a named vector of initial values for fitting}
}
\description{
Takes all the information needed and returns a cm-class object
}
\examples{
cat_function()


}
\keyword{custommodel}

