% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{fit}
\alias{fit}
\title{fits a model to data}
\usage{
fit(par = NULL, left = NULL, right = NULL, cm, method = "L-BFGS-B",
  control = list(maxit = 10000, factr = 10, trace = 0))
}
\arguments{
\item{par}{initial vector of parameters}

\item{left}{left margin}

\item{right}{left margin}

\item{cm}{an object of class custommodel}

\item{method}{method for optim}

\item{control}{control list}
}
\description{
a wrapper over optim function
}
\examples{
m2lL(par, cm, ode_solver = cm_ode_c, ...)


}
\keyword{likelihood}

