% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood.R
\name{m2lL}
\alias{m2lL}
\title{-2lnL}
\usage{
m2lL(par, cm, ode_solver = cm_ode_c, ...)
}
\arguments{
\item{par}{a vector of parameters}

\item{cm}{an object of class custommodel}

\item{ode_solver}{ode solver function (defaults to cm_ode_c)}
}
\description{
gets parameter and state values, return -2lnL
}
\examples{
m2lL(par, cm, ode_solver = cm_ode_c, ...)


}
\keyword{likelihood}

