% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/separate_data.R
\name{separate_data}
\alias{separate_data}
\title{Data separator}
\usage{
separate_data(dat)
}
\arguments{
\item{dat}{data.frame to be separated}
}
\description{
gets a data.frame, divides it into a list of data.frames so that each data.frame has unique experimental conditions
}
\examples{
separate_data(dat)


}
\keyword{likelihood}

